requirements.txt
apscheduler==3.8.1
dropbox==11.10.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.33.0
requests==2.26.0
python-dotenv==1.0.0
tzlocal==2.1
Flask==2.0.2
Werkzeug==2.2.2